refactor(account-api): split tests
